requirements.txt
flask==2.3.2
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2
marshmallow==3.19.0
flask-swagger-ui==4.11.1
apispec==6.0.2
openai==0.27.6